Relax nest's dependency on Query
